socketchat/chat/serializers/user_serializer.py
from rest_framework import serializers
from django.contrib.auth.models import User


class UserRequestSerializer(serializers.Serializer):
    username = serializers.CharField(required=True)
    password = serializers.CharField(required=True)


class UserResponseSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ("id", "username", "is_staff")
