socketchat/requirements.txt
ansible-base==2.10.8
apache-libcloud==3.2.0
argcomplete==1.8.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.8.0
black==23.9.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
channels==4.0.0
chardet==4.0.0
click==8.1.7
command-not-found==0.3
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
Django==4.2.5
django-cors-headers==4.2.0
django-csp==3.7
djangorestframework==3.14.0
dnspython==2.1.0
filelock==3.12.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
MarkupSafe==2.0.1
more-itertools==8.10.0
mypy-extensions==1.0.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psutil==5.9.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.11.0
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pyOpenSSL==23.2.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dotenv==1.0.0
pytz==2023.3.post1
pywinrm==0.3.0
PyYAML==5.4.1
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selinux==3.3
service-identity==23.1.0
shortuuid==1.0.11
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.4
systemd-python==234
tomli==2.0.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uuid==1.30
virtualenv==20.24.3
wadllib==1.3.6
xmltodict==0.12.0
zipp==1.0.0
zope.interface==6.0
